Refactor(pairplot): Have parent and child classes
